simple-text2sql/requirements.txt
langchain-core
pydantic
requests
python-dotenv
streamlit